requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
websockets==12.0
google-genai==0.3.0
python-dotenv==1.0.1
pydub==0.25.1
numpy==1.26.4
scipy==1.12.0
starlette==0.36.3
jinja2==3.1.3
python-multipart==0.0.9
